Reclassify runtime dependency (#33)
The runtime dependency is inactive by default today. The runtime framework version is not overridden unless UseCustomRuntimeVersion is true, which it is not. I believe this is from when we didn't have a .NET 6 TFM targeting SDK, and needed to use a custom framework version. That said, the usage is incorrect at the moment because:
- It doesn't take into account a ref pack that drifts from the runtime pack version.
- Microsoft.NETCore.App is no longer actually a package. Instead should be the x64 runtime pack version, as commonly used in other repos.

It also doesn't need to be a product dependency, so reclassifying as a toolset dep.